+++ fluentvoice/__init__.py
@@ -2,6 +2,6 @@
 Created by Nick Otmazgin
 """
 
-__version__ = "1.4.8"
+__version__ = "1.4.9"
 __author__ = "Nick Otmazgin"
 __license__ = "MIT"